requirements.txt
midjourney-api
pyserial
